Initial commit- Instagram clone for android using Firebase database.
Login/regiser activities using firebase auth completed.

MainActivity, the main feed for the app has current users that are added to the database appearing as a list view - needs to be changed to be more like instas design.

Working on the AddPostActivity to upload a image from the phone to the database so it can be viewed by other users.

Blank Activites created for the search, marketplace and users activies.